+++ apis/exp.ts
@@ -8,6 +8,7 @@ import {
 import API from './config';
 
 export interface ExperiencePayload {
+  id?: number;
   status: ExperienceStatus;
   experienceType: ExperienceType;
   formType: ExperienceFormType;
@@ -40,6 +41,12 @@ interface GetExperienceResponse {
   data?: Experience[];
 }
 
+interface GetExperienceByIdResponse {
+  httpStatus: number;
+  message: string;
+  data: ExperiencePayload;
+}
+
 export async function saveExperience(
   payload: ExperiencePayload
 ): Promise<SaveExperienceResponse> {
@@ -48,6 +55,27 @@ export async function saveExperience(
     startDate: new Date(payload.startDate),
     endDate: new Date(payload.endDate),
   });
+
+  return res.data;
+}
+
+export async function editExperience(
+  exp_id: number,
+  payload: ExperiencePayload
+): Promise<SaveExperienceResponse> {
+  const res = await API.patch<SaveExperienceResponse>(`/api/exp/${exp_id}`, {
+    ...payload,
+    startDate: new Date(payload.startDate),
+    endDate: new Date(payload.endDate),
+  });
+
+  return res.data;
+}
+
+export async function getExperienceById(
+  exp_id: number
+): Promise<GetExperienceByIdResponse> {
+  const res = await API.get<GetExperienceByIdResponse>(`/api/exp/${exp_id}`);
 
   return res.data;
 }
